MAINT: Move 'create_fracture_network ' to separate file
Needed to avoid cyclic imports when introducing a parent class for FractureNetworks
