+++ src/porepy/__init__.py
@@ -111,7 +111,7 @@ from porepy.utils.tangential_normal_projection import (
 from porepy.fracs.plane_fracture import PlaneFracture, create_elliptic_fracture
 from porepy.fracs.line_fracture import LineFracture
 from porepy.fracs.ellipse_fracture import EllipticFracture
-from porepy.fracs.fracture_network import create_fracture_network
+from porepy.fracs.fracture_network_creation import create_fracture_network
 
 
 # Wells
